use size_t in all functions needed

+++ programs/ft_bzero.c
@@ -6,15 +6,15 @@
 /*   By: abosc <abosc@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 10:26:21 by abosc             #+#    #+#             */
-/*   Updated: 2024/10/15 12:44:21 by abosc            ###   ########.fr       */
+/*   Updated: 2024/10/15 13:10:08 by abosc            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-// #include <string.h>
+#include <string.h>
 // #include <unistd.h>
 // #include <stdio.h>
 
-void	*ft_memset(void *s, unsigned int n)
+void	*ft_memset(void *s, size_t n)
 {
 	unsigned int	i;
 	unsigned char	*modifiable_s;
